# application/src/main/resources/static/openapi/canyoningApp-openapi.yml
openapi: 3.0.3
info:
  title: Swagger CanyoningApp - OpenAPI 3.0
  description: Swagger of canyoning app
  contact:
    email: delgadofernandez.luis@gmail.com
  version: 0.0.1
servers:
  - url: http://localhost:8080
  #- url: https://misbarrancosbdesarrollo.com:8443
...
paths:
...
  /canyons:
    post:
      tags:
        - Canyon
      summary: Create a new canyon
      description: Create a new canyon
      operationId: createCanyon
      requestBody:
        description: data from the canyon
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Canyon'
        required: false
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canyon'
        400:
          $ref: '#/components/responses/400BadRequest'
    get:
      tags:
        - Canyon
      summary: Get all canyon
      description: Get all canyon filtering by differents fields
      operationId: getCanyons
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: season
          in: query
          required: false
          schema:
            type: string
        - name: river
          in: query
          required: false
          schema:
            type: string
        - name: country
          in: query
          required: false
          schema:
            type: string
        - name: population
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Número de página (empezando desde 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Número de elementos por página
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Criterios de ordenación en el formato propiedad(,asc|desc). El orden por defecto es ascendente.
          required: false
          schema:
            type: string
            default: "canyonId,asc"
      x-spring-paginated: true
      responses:
        201:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimpleCanyon'
          headers:
            Link:
              $ref: '#/components/headers/Link'
            X-Page-Number:
              $ref: '#/components/headers/X-Page-Number'
            X-Total-Elements:
              $ref: '#/components/headers/X-Total-Elements'
            X-Total-Pages:
              $ref: '#/components/headers/X-Total-Pages'
        400:
          $ref: '#/components/responses/400BadRequest'
...
  /canyons/{canyonId}:
    get:
      tags:
        - Canyon
      summary: Get one canyon
      description: Get canyon by id
      operationId: getCanyon
      parameters:
        - name: canyonId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canyon'
        400:
          $ref: '#/components/responses/400BadRequest'
    put:
      tags:
        - Canyon
      summary: Update your canyon
      description: Update your canyon
      operationId: updateCanyon
      parameters:
        - name: canyonId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Canyon'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Canyon'
        400:
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - Canyon
      summary: Deletes a canyon
      description: delete a canyon
      operationId: deleteCanyon
      parameters:
        - name: canyonId
          in: path
          description: User id to delete
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        204:
          description: 'Succesfull delete'
        400:
          $ref: '#/components/responses/400BadRequest'
  /canyons/{canyonId}/download:
    get:
      tags:
        - Canyon
      summary: Download a canyon
      description: Download canyon by id and return a PDF file
      operationId: downloadCanyon
      parameters:
        - name: canyonId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - name: email
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successful operation
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/400BadRequest'
        '404':
          description: Canyon not found
        '500':
          description: Internal server error
...
  /activities:
    post:
      tags:
        - Activity
      summary: Create a new activity
      description: Create a new activity
      operationId: createActivity
      requestBody:
        description: data from the activity
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
        required: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        400:
          $ref: '#/components/responses/400BadRequest'
    get:
      tags:
        - Activity
      summary: Get all activity
      description: Get all activity filtering by differents fields
      operationId: getActivities
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: meetingPlace
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Número de página (empezando desde 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Número de elementos por página
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Criterios de ordenación en el formato propiedad(,asc|desc). El orden por defecto es ascendente.
          required: false
          schema:
            type: string
            default: "activityId,asc"
      x-spring-paginated: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimpleActivity'
          headers:
            Link:
              $ref: '#/components/headers/Link'
            X-Page-Number:
              $ref: '#/components/headers/X-Page-Number'
            X-Total-Elements:
              $ref: '#/components/headers/X-Total-Elements'
            X-Total-Pages:
              $ref: '#/components/headers/X-Total-Pages'
        400:
          $ref: '#/components/responses/400BadRequest'
  /activities/{activityId}:
    get:
      tags:
        - Activity
      summary: Get an activity
      description: Get activity by id
      operationId: getActivity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        400:
          $ref: '#/components/responses/400BadRequest'
    put:
      tags:
        - Activity
      summary: Update your activity
      description: Update your activity
      operationId: updateActivity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Activity'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        400:
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - Activity
      summary: Deletes a activity
      description: delete a activity
      operationId: deleteActivity
      parameters:
        - name: activityId
          in: path
          description: Activity id to delete
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        204:
          description: 'Succesfull delete'
        400:
          $ref: '#/components/responses/400BadRequest'
  /activities/{activityId}/share:
    get:
      tags:
        - Activity
      summary: Share your activity qr
      description: Share your activity qr
      operationId: shareActivity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
  /activities/{activityId}/share/{userId}:
    post:
      tags:
        - Activity
      summary: Insert an user on your activity
      description: Insert an user on your activity
      operationId: insertActivity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - name: userId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
...
  /images:
    post:
      tags:
        - Image
      summary: Post an image
      description: Post an image
      operationId: postImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                name:
                  type: string
                dir:
                  type: string
              required:
                - file
      responses:
        201:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
...
components:
  schemas:
...
    SimpleActivity:
      type: object
      properties:
        activityId:
          $ref: '#/components/schemas/Id'
        user:
          $ref: '#/components/schemas/SimpleUser'
        date:
          type: string
          format: date
        meetingPlace:
          type: string
        activityType:
          type: string
    Activity:
      allOf:
        - $ref: '#/components/schemas/SimpleActivity'
        - properties:
            capacity:
              type: integer
            meetingTime:
              type: string
            description:
              type: string
            qr:
              type: string
            participants:
              type: array
              items:
                $ref: '#/components/schemas/Id'
...
    SimpleCanyon:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        name:
          type: string
        season:
          type: string
        river:
          type: string
        population:
          type: string
    Canyon:
      allOf:
        - $ref: '#/components/schemas/SimpleCanyon'
        - properties:
            description:
              type: string
            croquis:
              type: string
            access:
              type: string
            approach:
              type: string
            descent:
              type: string
            _return:
              type: string
            scape:
              type: string
            location:
              $ref: '#/components/schemas/CanyonLocation'
            canyonRappeling:
              type: array
              items:
                $ref: '#/components/schemas/CanyonRappeling'
            canyonDescent:
              type: array
              items:
                $ref: '#/components/schemas/CanyonDescent'
            canyonSchedule:
              type: array
              items:
                $ref: '#/components/schemas/CanyonSchedule'
            canyonLink:
              type: array
              items:
                $ref: '#/components/schemas/CanyonLink'
            canyonProhibition:
              type: array
              items:
                $ref: '#/components/schemas/CanyonProhibition'
            canyonDifficulty:
              type: array
              items:
                $ref: '#/components/schemas/CanyonDifficulty'
            canyonCanyonNear:
              type: array
              items:
                $ref: '#/components/schemas/CanyonCanyonNear'
            canyonControlLevel:
              type: array
              items:
                $ref: '#/components/schemas/CanyonControlLevel'
    CanyonLocation:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        country:
          type: string
        population:
          type: string
        latitud:
          type: string
        longitud:
          type: string
        zone:
          type: string
    CanyonCanyonNear:
      type: object
      properties:
        principalCanyon:
          $ref: '#/components/schemas/Id'
        nearCanyon:
          type: string
    CanyonControlLevel:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        caudalLevel:
          type: string
        name:
          type: string
        controlPoint:
          type: string
    CanyonDifficulty:
      type: object
      properties:
        descentNumber:
          type: integer
        canyonId:
          $ref: '#/components/schemas/Id'
        difficultyDesc:
          type: string
    CanyonProhibition:
      type: object
      properties:
        canyonProhibitionId:
          $ref: '#/components/schemas/Id'
        canyonId:
          $ref: '#/components/schemas/Id'
        description:
          type: string
    CanyonLink:
      type: object
      properties:
        canyonLinkId:
          $ref: '#/components/schemas/Id'
        canyonId:
          $ref: '#/components/schemas/Id'
        link:
          type: string
        title:
          type: string
    CanyonSchedule:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        car:
          type: integer
        descentNumber:
          type: integer
        approach:
          type: string
        descent:
          type: string
        _return:
          type: string
    CanyonDescent:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        descentNumber:
          type: integer
        length:
          type: string
        slope:
          type: string
        rapelNum:
          type: string
        maxLength:
          type: string
        equipment:
          type: string
    CanyonRappeling:
      type: object
      properties:
        canyonId:
          $ref: '#/components/schemas/Id'
        descentNumber:
          type: integer
        step:
          type: string
        stepType:
          type: string
        length:
          type: string
        location:
          type: string
        description:
          type: string
...
    SimpleUser:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          example: theUser
        name:
          type: string
          example: John
        lastName:
          type: string
          example: James
        location:
          type: string
        description:
          type: string
        birthDay:
          type: string
          format: date
...
    Id:
      type: integer
      format: int64
    Error:
      type: object
      properties:
        errorCode:
          type: string
        description:
          type: string
        cause:
          type: string
  responses:
    400BadRequest:
      description: Bad Request. Check the parameters or body of the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    basic:
      type: http
      scheme: basic
  headers:
    Link:
      description: >-
        The Link header provides information about the pagination of the response.
      schema:
        type: string
    X-Page-Number:
      description: >-
        The current page number of the response.
      schema:
        type: integer
    X-Total-Elements:
      description: >-
        The total number of elements available in the response.
      schema:
        type: integer
    X-Total-Pages:
      description: >-
        The total number of pages available in the response.
      schema:
        type: integer
security:
  - basic: [ ]